refactor(RacingCar): refactor test class
Domain , Service 변화에 따른 test code 수정
